Add a simple experiments runner.
This runs 4 experiments: cpctplus; cpctplusdyndist; mf; and "reverse mf". The
latter inverts the ranking order of repair sequences i.e. it guarantees to
always pick from the worst (and not the best) repair sequences.

In order for the runner to work, we have to slightly patch lrpar so that it
prints out how much time it spent in recovery. That then allows the runner to do
its work and record output in CSV output files.