Rewrite the results processor: generate confidence intervals; calculate costs.
This is a very big change. We now run each error recovery algorithm on each
source file PEXECS (=30) times. When generating results, we then use this
(huge!) amount of data to bootstrap and produce 99% confidence intervals. As a
side bonus, we also now collect the cost of each repair sequence found, which
allows us to produce statistics on the mean cost of repair sequences.