Add a static Java parser.
Rather than compiling the Java grammar every time we parse a file, this uses
lrpar's builder feature to statically compile the Java parser in. This should
make the experiment quite a bit faster to run.

This means that we need to fairly heavily change the build infrastructure.
Whilst here, add a binary to test out the panic mode recoverer.