Don't count input skipped because of incomplete error recovery.
Previously "skipped input" consisted of input skipped by Delete repairs, as well
as input skipped at the end of a file because we couldn't repair it. In
retrospect, this second thing has very little relation to the first. What we're
really interested in is how many Delete repairs there are, so delete the second
thing from our measurement (noting that the second thing is captured
well-enough-for-our-purposes in the "failure rate" metric).