Rethink the way decimal places are reported.
This is a small commit, but it's taken over 2 weeks to get right. The basic
problem is that the new data we have has a) extremely small times for panic mode
b) often very small confidence intervals. Working out how to present these has
been a real challenge, because with the old technique gave us things like
"1.23%+/-0.000" which makes little sense.

Thanks to a suggestion from Carl Friedrich Bolz-Tereick, we now print such tiny
confidence intervals as "1.23%+/-<0.001" i.e. saying that the confidence
interval is less than 0.001%. This is much more intuitive and makes the output
from this script much easier to read.