Use a breadth-first search for Corchuelo.
The Corchuelo paper is fairly confusing in this regard: it says it uses a
breadth-first search but then suggests its queue is a list to which things are
appended (which, by definition, isn't a breadth-first queue). In my first
implementation I assumed the "append" factor outweighed the former but, on
reflection, I think it's better to assume that "breadth-first" is more likely to
reflect their intent. It allows us to use Cerecke's data-structure, which gives
a useful performance improvement.