Include non-zero numbers of error locations even when the pexec failed.
In the error locations histogram, it's fine if a pexec did not succeed in the
recovery budget: it still gives us a lower bound of the number of errors in the
file. However, it can happen for the slower recoverers (e.g. CPCT+rev) that no
error locations are found before the timeout, at which point we get confusing
bars in the histogram suggesting that there are some files with no error
locations (when we know that every file in the set has at least one error
location), so we exclude that case.